catchpy/requirements/base.txt
Django~=4.2
iso8601~=2.0.0
jsonschema==4.18.4
psycopg>=3.1.8
PyJWT==2.8.0
PyLD==2.0.3
python-dotenv==1.0.0
requests~=2.31.0
django-log-request-id==2.1.0
django-cors-headers~=4.2.0
